Use built-in to determine ctrlr mode. Fix #3.
The previous implementation tried to deduce in which mode the controller
was by looking at the value of the 'CE/CR' bits. While this works on all
tested real controllers, the virtual robot controllers in Roboguide don't
seem to set those bits, resulting in the wrong controller mode being
reported.

The IN_AUTO_MODE() built-in was found to work correctly, both in simulation
as well as on real controllers.