driver: add KAREL Vars config to state & relay.
These structures allow runtime configuration of various important settings
and limits. See the wiki for more information on how to access these
structures.

This fixes #2, fixes #8 and fixes #71.

---

Backport from driver_ng. The 'KAREL Vars' interface will be replaced
with a proper custom Form and menu in a later commit.