driver: add Makefile and robot.ini to build driver.
This allows building the driver from the command line, in stead of
using Roboguide. The Makefile by default builds for V7.70 systems,
but this can be overriden by setting the SUPPORT_VER variable. Example:

  make SUPPORT_VER=V8.20-1

This would build for V8.20 systems.

The included robot.ini file seems to be the absolute minimum needed
to succesfully compile the Karel sources. P-code files will not be
copied to any Roboguide workcells, but are written to the cwd.

Note: ktrans.exe is expected to be on the path.