moveit_cfg: make warehouse db location user configurable. Fix #142.
These changes prevent the MoveIt mongodb wrapper script from running unless
the user explicitly requests it to be enabled by setting `db:=true` when
starting `demo.launch` or `moveit_planning_execution.launch`.

Additionally, when starting the wrapper script from locations that are
non-writable for the current user, the `db_path` launch file argument may
be used to specify an alternate location for the database.

Specifying an alternative location is required when using MoveIt Configuration
packages that have been released as Debians (as those are typically installed
into non-writable directories).

Essentially a backport of the fix in ros-planning/moveit_setup_assistant#103.