moveit_cfg: load xacros instead of urdfs.
Instead of loading the '.urdf' top-levels, this change makes all MoveIt config
pkgs load the top-level '.xacro'. Even though users are expected to only use
the provided MoveIt config pkgs as examples or for initial exploratory
planning, they should still be easily editable and updateable. By loading the
top-level xacro directly, users avoid having to run the xacro through xacro.py,
allowing them to skip that step.

In addition, this change updates demo.launch and
moveit_planning_execution.launch to use the 'load_*.launch' files provided by
the robot support pkgs. This is more in-line with the recommended use of
launch files in support pkgs, as it uses the 'public api' of the support pkg
(as described in [1]), instead of directly accessing the xacro.

[1] http://wiki.ros.org/Industrial/Tutorials/WorkingWithRosIndustrialRobotSupportPackages