moveit_cfg: change default planner to RRTConnect. Fix #179.
MoveIt previously defaulted to LBKPIECE which does not always result
in consistent planning performance or logical/efficient/sensable
trajectories. RRTConnect seems to perform better, hence it is set
as the default.

Note: this is a default planner, which is used only when planning
requests don't specify one as part of the request.