moveit_cfgs: add roslaunch testing. For ros-industrial/fanuc#104.
This adds basic roslaunch testing to all MoveIt configs in this repository.

Only `demo.launch` and `moveit_planning_execution.launch` are tested for now,
and the latter only in 'non-sim' mode due to a bug in roslaunch that causes
it to not handle the optional argument `robot_ip` properly (as it has no default).

`demo.launch` is tested directly, `moveit_planning_execution.launch` is tested
through a small wrapper that supplies values for the required arguments.

Finally, `roslaunch` is added as a `test_depend` to all package manifests.