support: add 'flange' frame.
The 'flange' frame is intended to function as an 'attachment point' for
EEF xacros. In order to easily swap robot models, it is beneficial for
such attachment points -- typically the last link in a robot xacro -- to
have a consistent orientation.

'flange' is defined as always having x+ pointing out of the (physical)
flange link. For robots that happen to have the link-local coordinate
frame of their last link in their zero pose already aligned with REP-103
conventions (x+ forward, y+ left, z+ up), 'flange' will be an identity
transform. In other cases, 'flange' includes appropriate rotations to
orient x+ such that it again follows the above described convention.

Note that the above does not mean that 'flange' will *always* be world
aligned, only that x+ will always point away from the (physical) flange /
last link of the robot model. Rotations and translations will impact
'flange's pose just as they do other robot links.

It's sibling frame, 'tool0', is similar, but this frame is *always*
located and oriented such that it corresponds to whatever is configured on
the robot controller as an 'all zeros' tool frame.

Both 'tool0' and 'flange' are to be considered immutable frames, and are
not intended to be changed or updated by consumers of xacro macros
including them.