moveit_cfgs: disable db tests for now.
The Python infrastructure in warehouse_ros was removed in ros-planning/warehouse_ros#32 and will not come back. This essentially breaks the DB functionality of all MoveIt configuration packages, including the ones in ros-industrial/fanuc.

For now the roslaunch tests that attempt to enable the DB will be disabled, as this is not an issue with those packages, but with warehouse_ros/MoveIt.

If/when that is resolved, the tests can be enabled again.